+++ CHANGELOG.md
@@ -14,6 +14,10 @@ Compatibility:
 * Implement more `Ripper` methods as no-ops (#1694).
 * Implemented `rb_enc_sprintf` (#1702).
 
+Changes:
+
+* Interactive sources (like the GraalVM polyglot shell) now all share the same binding (#1695).
+
 # 20.0.0 beta 1
 
 Bug fixes: